backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.15.0
pillow==8.4.0
numpy==1.21.0
python-dotenv==0.19.0
gunicorn==20.1.0